refactored subscription controller and its services
