+++ package.json
@@ -8,7 +8,8 @@
         "eslint:check": "eslint src/",
         "format:fix": "prettier --write src/",
         "format:check": "prettier --check src/",
-        "prepare": "husky install"
+        "prepare": "husky install",
+        "test": "jest"
     },
     "dependencies": {
         "body-parser": "^1.20.0",
@@ -17,9 +18,7 @@
         "debug": "~2.6.9",
         "dotenv": "^16.0.1",
         "express": "~4.16.1",
-        "express-form-data": "^2.0.18",
-        "http-errors": "~1.6.3",
-        "jade": "~1.11.0",
+        "http-errors": "^2.0.0",
         "morgan": "~1.9.1",
         "nodemailer": "^6.7.7"
     },
@@ -29,6 +28,7 @@
         "eslint-plugin-import": "^2.26.0",
         "eslint-plugin-prettier": "^4.2.1",
         "husky": "^8.0.1",
+        "jest": "^29.0.1",
         "prettier": "^2.7.1"
     }
 }
