+++ src/bin/www
@@ -12,6 +12,7 @@
  server.listen(port);
  server.on('error', onError);
  server.on('listening', onListening);
+ console.log(`Server is running on port ${port}`);
  
  function normalizePort(val) {
   const port = parseInt(val, 10);
