+++ src/data/emails.json
@@ -1,1 +1,1 @@
-{ "emails": [] }
+{ "emails": [""] }
