+++ src/services/emailService.js
@@ -1,5 +1,8 @@
 require('dotenv').config();
 const nodemailer = require('nodemailer');
+const createError = require('http-errors');
+const getRate = require('./rateService');
+const { readFile } = require('./dbService');
 
 const transporter = nodemailer.createTransport({
     host: process.env.EMAIL_HOST,
@@ -11,4 +14,17 @@ const transporter = nodemailer.createTransport({
     },
 });
 
-module.exports = { transporter };
+module.exports = async () => {
+    try {
+        const emails = await readFile();
+        const mailOptions = {
+            from: process.env.EMAIL_NAME,
+            to: emails,
+            subject: 'Bitcoin price',
+            text: `Bitcoin price is ${await getRate()} UAH`,
+        };
+        await transporter.sendMail(mailOptions);
+    } catch (err) {
+        throw createError(400, err.message);
+    }
+};
